feat: add base for handling per-output to base-layout merging
part of issue #401.

My idea would be to have a hashmap of layouts keyed by the output name.

Downside is that we're essentially creating clones of the base layout
for every output with configuration deviations but I would assume there
aren't so many that there would be an issue with this.

We can then have a simple layout getter that gets either the base layout
or output specific layout if there are deviations from the base.